# information-srv/client/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Contain Generic Information.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Information API
  version: "1.0"
servers:
- url: /
paths:
...
  /internal/product:
    get:
      description: Retrieve a complete list of all products.
      parameters:
      - description: Id
        explode: true
        in: query
        name: id
        required: false
        schema:
          type: string
        style: form
      - description: Name
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      - description: AccountTypeCode
        explode: true
        in: query
        name: accountTypeCode
        required: false
        schema:
          type: string
        style: form
      - description: AccountType
        explode: true
        in: query
        name: accountType
        required: false
        schema:
          type: string
        style: form
      - description: ExchangeMarketId
        explode: true
        in: query
        name: exchangeMarketId
        required: false
        schema:
          type: string
        style: form
      - description: Suffix
        explode: true
        in: query
        name: suffix
        required: false
        schema:
          type: string
        style: form
      - description: TransactionType
        explode: true
        in: query
        name: transactionType
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_product_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Products
      tags:
      - Product
...
  /secure/product:
    get:
      description: Retrieve a complete list of all products.
      parameters:
      - description: Id
        explode: true
        in: query
        name: id
        required: false
        schema:
          type: string
        style: form
      - description: Name
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      - description: AccountTypeCode
        explode: true
        in: query
        name: accountTypeCode
        required: false
        schema:
          type: string
        style: form
      - description: AccountType
        explode: true
        in: query
        name: accountType
        required: false
        schema:
          type: string
        style: form
      - description: ExchangeMarketId
        explode: true
        in: query
        name: exchangeMarketId
        required: false
        schema:
          type: string
        style: form
      - description: Suffix
        explode: true
        in: query
        name: suffix
        required: false
        schema:
          type: string
        style: form
      - description: TransactionType
        explode: true
        in: query
        name: transactionType
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_internal_product_get_200_response'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pkg.Response'
          description: Internal Server Error
      summary: Get All Products
      tags:
      - Product
components:
  schemas:
...
    pkg.Response:
      example:
        detail: detail
        title: title
        status: 6
      properties:
        detail:
          type: string
        status:
          type: integer
        title:
          type: string
      type: object
    product.Product:
      example:
        account_type_code: account_type_code
        account_type: account_type
        name: name
        id: id
        suffix: suffix
        transaction_type: transaction_type
        exchange_market_id: exchange_market_id
      properties:
        account_type:
          type: string
        account_type_code:
          type: string
        exchange_market_id:
          type: string
        id:
          type: string
        name:
          type: string
        suffix:
          type: string
        transaction_type:
          type: string
      type: object
...
    _internal_product_get_200_response:
      allOf:
      - $ref: '#/components/schemas/pkg.Response'
      - properties:
          data:
            items:
              $ref: '#/components/schemas/product.Product'
            type: array
        type: object
      example:
        data:
        - account_type_code: account_type_code
          account_type: account_type
          name: name
          id: id
          suffix: suffix
          transaction_type: transaction_type
          exchange_market_id: exchange_market_id
        - account_type_code: account_type_code
          account_type: account_type
          name: name
          id: id
          suffix: suffix
          transaction_type: transaction_type
          exchange_market_id: exchange_market_id
        detail: detail
        title: title
        status: 0